requirements.txt
influxdb==5.3.1
python-dateutil==2.8.1
requests==2.25.0
urllib3==1.26.2
websocket-client==0.57.0
